refactor(stage1): move pure text helpers to woeai.publications.textutils
Candidate A, slice 2. Moves the pure-string helpers and shared constants
that have no data/network dependency out of update-publications-from-zotero.py:

- woeai/publications/textutils.py: ZoteroError, METRIC_LABELS,
  GROUP_LEADER_AUTHOR_NAMES, CORRESPONDING_AUTHOR_LABELS,
  PINYIN_SURNAMES, CHINESE_INITIALS, JOURNAL_INITIALISM_OVERRIDES, plus
  extract_year, parse_date, normalize_title, normalize_doi, contains_cjk,
  slug, chinese_initialism, rst_escape, strip_bib_html, split_extra_tokens.
- The script re-exports them so its public surface (tests access it via
  self.updater.<name>) stays identical - 54 tests green with no test changes.

Script shrinks by 184 lines; these names are the shared base layer for the
remaining author/render/data slices.
